+++ completions/tspin.bash
@@ -25,7 +25,7 @@ _tspin() {
 
     case "${cmd}" in
         tspin)
-            opts="-f -p -e -h -V --follow --print --config-path --exec --highlight --enable --disable --extras --disable-builtin-keywords --pager --generate-bash-completions --generate-fish-completions --generate-zsh-completions --help --version"
+            opts="-f -p -e -h -V --follow --print --config-path --exec --highlight --enable --disable --extras --disable-builtin-keywords --pager --generate-bash-completions --generate-fish-completions --generate-zsh-completions --generate-default-theme --help --version"
             if [[ ${cur} == -* || ${COMP_CWORD} -eq 1 ]] ; then
                 COMPREPLY=( $(compgen -W "${opts}" -- "${cur}") )
                 return 0